+++ src/main/java/com/example/d3nserver/common/exception/CustomException.java
@@ -1,6 +1,5 @@
 package com.example.d3nserver.common.exception;
 
-import com.example.d3nserver.common.dto.ErrorCode;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
